+++ packages/perseus/src/widgets/grapher/grapher.tsx
@@ -18,7 +18,7 @@ import KhanColors from "../../util/colors";
 import {getInteractiveBoxFromSizeClass} from "../../util/sizing-utils";
 /* Graphie and relevant components. */
 /* Mixins. */
-import {getPromptJSON as _getPromptJSON} from "../../widget-ai-utils/grapher/prompt-utils";
+import {getPromptJSON as _getPromptJSON} from "../../widget-ai-utils/grapher/grapher-ai-utils";
 
 import scoreGrapher from "./score-grapher";
 import {
@@ -40,7 +40,7 @@ import type {
     PerseusGrapherRubric,
     PerseusGrapherUserInput,
 } from "../../validation.types";
-import type {GrapherPromptJSON} from "../../widget-ai-utils/grapher/prompt-utils";
+import type {GrapherPromptJSON} from "../../widget-ai-utils/grapher/grapher-ai-utils";
 import type {PropsFor} from "@khanacademy/wonder-blocks-core";
 
 // @ts-expect-error - TS2339 - Property 'MovablePoint' does not exist on type 'typeof Graphie'.
